Change capacity default value, add hash method
Capacity represents the number of buckets or positions to store data in in our hashmap, which I'll start at a value of 16.
The hash method is the method that calculates a hash given a key value, and the value it returns will be used as the index in the hashmap to reach one of these buckets of data. I used a hash function that was defined by the lesson and added the action of dividing the calculation by the capacity value. This capacity value can change later when more space is needed, and when that is required then all keys will be revaluated to reposition data.