Add starting variables of loadFactor and capacity
Instead of using a class and a constructor, we set these variables up with default values in the parameters section so our instances automatically start with these variables and values.
Again I picked 16 for capacity because why not, and .75 for loadFactor because I saw it referenced in the assignment directions.