Add a modulo operation inside the hash method
Typically we'd use modulo on the final result of the iteration building a variable holding the accumulated hash value, but the assignment has warned us to not do that because we could potentially receive an argument that would result in unwanted outcomes if we didn't calculate the modulo inside the loop that accumulates that sum hash value.
If this is incorrect then I'll come back to it but for now I'll leave it as such and I'll add a comment to remind myself for later.