Add method requirement comments, add get method
I've gone ahead and added the instructions/requirements for each method above the method declaration, this way I reduce the number of times I have to look at the project assignment page.
I've added the get method, which when the user gives a key we return the value stored via hashing that key.