Add the has method
This method needs to return if a key is in the hash map.
What I found weird is the wording of this methods requirements because we can use the given condition that triggers a response if the index calculated is greater than our range or if it's less than zero to return false and then anything else return true.
However I thought just because a key can be calculated to a index of our hash map that does not mean there is a value there. So I've gone ahead and include a check to look if there is a value when the key argument does return an appropriate index value.