Change remove method operation, add clear method
I realized while adding the clear method  logic that the assignment didn't want us to insert a new linkedList when clearing or removing but to set the value as a clean state aka how it was before any value was set. Which would be === undefined. Therefore I've changed the remove method to use the delete operator on specific hash map indexes accordingly and the clear method iterated through all buckets and invokes the operator as well.