Change set, values, keys, entries methods
The set method was creating a new nodeInstance instance, which wasn't needed because the append method from linked_list already does that, so I change the code to pass the appropriate values as arguments to the append invocation.
The keys, values, and entries methods used concat incorrectly so I changed them to use it correctly. Concat does not modify the array you are somewhat pushing another arrays elements into, it returns that altered array though so I've overridden the old array when this happens with that appropriate modified array.
Also in the set method, I had a condition to execute code when bucket did not equate to undefined, previously I used null, but I forgot that how we use the array from method to create the buckets array we instead give them each a value of null until we change them in the set method when that is used, so I changed the condition to correctly execute its body of code when an element is equal to null.