Add braces and remove unnecessary curly
The entries method is supposed to return an array of arrays, each of those sub arrays is made up of the key then its value separated by a comma. Before these changes the array did not have sub array just string pairs and and unnecessary curly brace at the end.