Change clear method
The logic I used before did not work; I attempted to use the delete operator but this was wrong for 2 reasons, first I was using forEach which is for side effects really and not modifying an arrays elements, and secondly the delete operator would have left an empty bucket when we needed null to be the new result.  The new code works as intended.