Add containsKey method, edit leading comments
I've added a new method that allows for the ability to check if a node exists in the linked list with a specific key, and not using the value.
I've also included changes to the comments above methods as I previously changed the methods to use keys in addition to values.