Add condition inside containsKey
Before these changes the method had a condition that checked if the head of the linkedList was === null, but there is a possibility that the linkedlist === null as well so added a check for it. It's really not a linked list if ===null it's an empty bucket but it's what we need here.