Change hash divider to capacity, add load methods
The hash function used a modulo operator to create hash values that were correlated to how many buckets there are in our hashmap, but the variable used was how many buckets there were and capacity is better as we will need to call this has function dynamically later when extending the bucket count/indexes in the buckets array.
Added loadCheck and doubleCapacity methods, loadCheck will be used when adding a value to the hashmap and if it returns true then doubleCapacity will be invoked which results in not only doubling the capacity but reorganizing the stored nodes into the new doubled capacity array.