Add ability to overwrite nodes values in set
Before this change I did not have the logic written to replace the value in a preexisting node with the new value entered into set when obviously a pre existing node collided with a key value pair passed to set.
This change required a fix to a method in linked_list.js which I had forgotten to fix so this was an excellent time for this functionality to be added.