Change condition inside set
This method has logic where it overwrites the value at a node when the same key is entered as an argument to the set method. Prior to these changes the method did not actually iterate through a linked list to find where the duplicate key was located and was incorrectly writing the value into the parent object in the bucket at a new property of value.